fraudguard-ai-backend/requirements_light.txt
fastapi==0.116.1
uvicorn==0.35.0
torch==2.7.1
torchvision==0.22.1
pillow==11.3.0
python-multipart==0.0.20